+++ packages/docs/src/App.tsx
@@ -1,27 +1,15 @@
 import { BrowserRouter } from 'react-router-dom'
 
 import { Provider } from '@siakit/core'
-import { DialogProvider } from '@siakit/dialog'
-import { LoadingProvider } from '@siakit/loading'
-import { ToastProvider } from '@siakit/toast'
 
-import { DefaultLayout } from './pages/_layouts/Default'
-import { AppRoutes } from './Routes/AppRoutes'
+import { AppRoutes } from './routes/AppRoutes'
 
 export function App() {
   return (
     <Provider>
-      <ToastProvider>
-        <LoadingProvider>
-          <DialogProvider>
-            <BrowserRouter>
-              <DefaultLayout>
-                <AppRoutes />
-              </DefaultLayout>
-            </BrowserRouter>
-          </DialogProvider>
-        </LoadingProvider>
-      </ToastProvider>
+      <BrowserRouter>
+        <AppRoutes />
+      </BrowserRouter>
     </Provider>
   )
 }