+++ packages/docs/src/pages/Home.tsx
@@ -1,0 +1,9 @@
+import { Flex } from '@siakit/layout'
+
+export function Home() {
+  return (
+    <Flex flex margin="0 auto" maxWidth={992}>
+      <p>Home</p>
+    </Flex>
+  )
+}