+++ packages/docs/src/pages/_layouts/Default/index.tsx
@@ -1,0 +1,19 @@
+import { Outlet } from 'react-router-dom'
+
+import { Flex } from '@siakit/layout'
+
+import { Sidebar } from './Sidebar'
+
+export function DefaultLayout() {
+  return (
+    <Flex overflow>
+      <Sidebar />
+
+      <Outlet />
+
+      <Flex width={320} padding>
+        aside
+      </Flex>
+    </Flex>
+  )
+}