+++ packages/form-components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @siakit/form-components
 
+## 0.0.28
+
+### Patch Changes
+
+- fix react hook form components
+
 ## 0.0.27
 
 ### Patch Changes